Added script for deletion Changes in checking the deletion permissions

+++ observation_data/observation_management.py
@@ -3,7 +3,7 @@ import logging
 from django.core.exceptions import BadRequest
 from nc_py_api import NextcloudException
 
-from accounts.models import ObservatoryUser, UserPermission
+from accounts.models import ObservatoryUser
 from nextcloud.nextcloud_manager import (
     generate_observation_path,
 )
@@ -14,14 +14,12 @@ import nextcloud.nextcloud_manager as nm
 logger = logging.getLogger(__name__)
 
 
-def delete_observation(user: ObservatoryUser, observation_id: int):
+def delete_observation(observation_id: int):
     """
     This function can be used to delete an observation from the database (and if existent from the nextcloud).
 
-    :param user: The user who tries to delete the observation.
     :param observation_id: The id of the observation.
     :raises ValueError: If no such observation exists.
-    :raises PermissionError: If the user does not have permission to delete the observation.
     :raises BadRequest: If the observation is already marked for deletion.
     """
 
@@ -29,16 +27,6 @@ def delete_observation(user: ObservatoryUser, observation_id: int):
         obs = AbstractObservation.objects.get(id=observation_id)
     except AbstractObservation.DoesNotExist:
         raise ValueError(f"Could not find observation with id {observation_id}")
-
-    if not user == obs.user and not user.has_perm(
-        UserPermission.CAN_DELETE_ALL_OBSERVATIONS
-    ):
-        logger.info(
-            f"User {user.get_username()} does not have permission to delete observation {observation_id}."
-        )
-        raise PermissionError(
-            f"User {user.get_username()} does not have permission to delete observation {observation_id}."
-        )
 
     if obs.project_status == ObservationStatus.PENDING_DELETION:
         raise BadRequest(f"Observation {obs.id} is already marked for deletion.")
@@ -56,10 +44,12 @@ def delete_observation(user: ObservatoryUser, observation_id: int):
     logger.info(f"Observation {obs.id} deleted successfully from database.")
 
 
-def process_pending_deletion():
+def process_pending_deletion(delete_user: bool = True):
     """
     Deletes all observations with status PENDING_DELETE from database and if existent from nextcloud.
     Also deletes all users with status
+
+    :param delete_user: If true, deletes all users with status PENDING_DELETION. Not wanted for deletion command
     """
     try:
         nm.initialize_connection()
@@ -83,5 +73,6 @@ def process_pending_deletion():
         logger.info(f"Observation {obs.id} deleted successfully from database.")
     logger.info("All observations with status PENDING_DELETE deleted successfully.")
 
-    for user in ObservatoryUser.objects.filter(deletion_pending=True):
-        user.delete()
+    if delete_user:
+        for user in ObservatoryUser.objects.filter(deletion_pending=True):
+            user.delete()
